Generalized quoting; LIT-WORD!/LIT-PATH! change, LIT
This takes the next step toward using generalized quoting, by accepting
the apostrophe as the character for it...and eliminating the special
cases of LIT-WORD!, LIT-PATH!, and LIT-BAR!.

Consequently, LIT-WORD! and LIT-PATH! are no longer DATATYPE!, but are
QUOTED! datatypes of WORD! and PATH! respectively.  This draws
attention to some longstanding problems of the usability of datatype,
which are being discussed (and needed to be addressed anyway).  In
the meantime, definitions for LIT-WORD! and LIT-PATH! are still present
along with type tests LIT-WORD? and LIT-PATH?, which in many cases
should operate as before.

One major point of compatibility change is that the ANY-WORD! and
ANY-PATH! typesets no longer contain LIT-WORD! or LIT-PATH!.  It's
possible to use quoting tools like `any-path? dequote value` to ask
if a value's unquoted form is an ANY-PATH!.  Or to use something like
`match [any-path! 'any-path!] value` to get 1 or 2 levels of quoting
to match a value.  However, these are all new and will take some amount
of getting used to.

The new datatype's name is QUOTED! instead of LITERAL!.  In accordance
with being more correct about the usage of what "literal" means in
programming--e.g. 1 is a "literal integer" and '1 is a "quoted integer"
this transitions the operator previously known as QUOTE to now be LIT
(or LITERAL).  QUOTE is temporarily disabled, but will become the
operation that adds a quote level to a value:

    >> x: 1

    >> quote x
    == '1

    >> lit x
    == x