Kill off SPEC/BODY form of object specification
The appearance of a need for a non-evaluative form of object creation
led to Ren-C having a "spec/body" form of MAKE OBJECT!, e.g.:

     make object! [[x y z] [x: <whatever> z: (1 + 2)]]

Being non-evaluative, this made it possible for z to literally hold
the GROUP! (1 + 2), and not be evaluated as 3.  The existence of the
spec was a way of working around the fact that a non-evaluative
notation cannot hold nulls...so the existence of y in the spec without
a corresponding pair in the body block would indicate that it was
supposed to be null in the final object.

A much simpler solution arose with generalized quoting.  Any values
that wish to be escaped can be...including escaped values.  NULLs
are escaped as simply an apostrophe with nothing after it.

This dials back some of the complexity that was introduced to support
this speculative feature, by eliminating the object spec block format
and using generalized quoting.  To be able to run the MAKE for an
object that contained voids in it, this also makes it so that a direct
quoted void literal can be assigned via SET-WORD!...even though void
assignments usually cause errors.

Also breaks out frame and error creation into their own MAKE_XXX
functions, since the %types.r dispatch allows for it.