Remove support for //-style comments
Trial of the new style for comments indicated they weren't all that
satisfying.

Instead, changes related to the NewPath proposal introducing immutable
paths suggest that `/a`, `a/`, and `/a/` all be paths...where implied
slots are empty and have blank in them ([_ a], [a _], and [_ a _]).

In such a world, it makes sense that when the block [_ _ _] becomes
represented as a path, it should display as `//`.  This purpose seems
to fit into the pattern of the language and its dialecting powers
in a more useful way.