Restructure parameter processing in paramlist creation
The internal code that analyzes function specs is now fairly old, and
had some structure that made it hard to adapt easily to changes in the
function spec operation (e.g. QUOTED! and PATH! vs. LIT-WORD! and
REFINEMENT!)  This separates out a reorganizational step so it can be
tested independently.