Streamline Data Stack Macros and Callsites
When the data stack was being revamped in early Ren-C, there was a
desire to avoid a PUSH(value) macro which would wastefully be copying
value bits instead of doing optimized initialization.  e.g.

    DS_PUSH(BLANK_VALUE); // copies 4x pointers, does binding checks...

    DS_PUSH_TRASH;
    Init_Blank(DS_TOP); // writes 1x pointer-sized header, no movement

Over time the conventions have become more certain and clear of how
things like END, TRASH, Unreadable BLANK!, etc. work.  In the face of
that certainty it seems like the best answer is to make DS_PUSH a macro
that takes no arguments, to help suggest it takes action *and* does
not move any values.  It's the caller's responsibility to use the
initialized cell.

There is enough checking in the debug build that you're not doing this
incorrectly that erring on the side of cleanness and brevity is better.
If a Move_Value() or Derelativize() is involved, it's fine to show
that at the call site.  (Most callers directly initialize...)