Bugfix for `<with> return` in FUNC specs
The reorganization of the spec processing code to make it easier to
work with non-ANY-WORD!s (like QUOTED! and PATH!) broke the case where
the `<with>` that was used to suppress the creation of definitional
returns for that function.

Includes some stray `//`-comment cleanup.

Disables brew in old OS X, because it keeps hanging.