Disallow CELL_KIND() on RELVAL*
One premise of safety in in-situ escaping is to stop making VAL_TYPE()
calls on a value when it is being "viewed" as its underlying payload
form.  This allows something like a MOLD of a quoted in-situ cell to
have one phase where it looks at that cell as quoted (and outputs an
apostrophe), then turns around and passes that same pointer--now a
view as the quoted type--into molding.  This prevents the infinite
loop of seeing it as quoting a quoted cell indefinitely.

This adds a test for the other direction--when a cell is not being
"viewed" as its underlying type, to disallow the modulo 64 trick for
getting its type.  This prevents someone from asking a quoted cell that
has been escaped into a new cell its kind, and getting REB_QUOTED mod
64, which is the sign that it is missing a VAL_UNESCAPED() call.