WORDS OF => PARAMETERS OF, fix for partial refinements
Partial refinement specialization can cause a reordering of the
effective parameters of an action.  This was not being taken into
account when listing parameters...which did not affect the execution,
but things that were based on reflection (HELP, R3-Alpha APPLY in
usermode) were broken.

This addresses the issue, and switches the reflector from WORDS OF
to be named PARAMETERS OF (especially with 'FOO being a QUOTED! and
/FOO being a PATH!, they will no longer be words.)

Also gets rid of the ARITY-OF, INFIX?, NFIX? and PREFIX? functions.
It does so despite those functions being much closer to working than
in any previous Rebol, due to the sketchy semantics of trying to gloss
over the information in PARAMETERS OF.  Further explanation:

https://forum.rebol.info/t/arity-of-then-and-now/1000