Expand MATCH dialect, add /ELSE, cut a feature to MATCH2
This brings some new powers to MATCH, where PATH!s are used to AND
together the individual rules, while BLOCK!s OR them together.  It is
more PARSE-like in being able to expand a variable into its rule, and
then recursively expand that rule.

EITHER-TEST is replaced by giving MATCH an /ELSE refinement.  This fits
better with calling this the "MATCH dialect".

The complexities of interactions with variadics need to be sorted out
for the frame-building "magic" MATCH feature, so that has been moved
to MATCH2 for the moment...pending further work.