EITHER-TEST soft quotes test, `<opt>` in MATCH dialect
With universal quoting, quoted types came to be the types of quoted
values at that level of quoting.  This was then turned around to be
used as a general form of function spec type checking.  e.g.:

    foo: func [x [word! lit-word!]] [...]

Became:

    foo: func [x [word! 'word!]] [...]

This is interesting and makes sense, and can be generalized to 'group!
or ''integer! or '''any-array!, etc.

It applies to MATCH too:

    >> match ['integer!] first ['2]
    == '2

But unlike type specs in functions, many uses of MATCH did not have the
types in a block.  This causes problems since the lit-marks would be
removed by the evaluator:

    >> match 'integer! first ['2]
    ;-- null

That's just the same as if you had written MATCH INTEGER! FIRST ['2].

Technically speaking MATCH soft quoted its first argument already, so
it didn't *have* to take that interpretation.  But it was trying to
emulate compatibility with EITHER-TEST that did not soft quote...the
reasons it did soft quote were for its own unrelated purposes.

The options here would be either to mandate a BLOCK! on the type or
to endorse a soft-quoting convention.  Urging the need to go with
soft quoting is the importance of being able to match what were
LIT-WORD!s in PARSE easily.  Previously one could say:

    >> did parse ['a 'b 'c'] [some lit-word!]
    == #[true]

But since 'WORD! would mean literally the word WORD!, it's needed to
throw in an PARSE keyword that means specifically interpreting things
as a matched type:

    >> did parse ['a 'b 'c'] [some match 'word!]
    == #[true]

That's not so bad, considering that it gained a lot of generality in
the process.  But if it's requiring a BLOCK! just in order to keep
consistent semantics to what MATCH in the normal evaluator would do,
that would be obscuring.

This biases it to wanting to say that MATCH and EITHER-TEST soft quote
their first argument, hence `match 'integer! first ['1]` would give
'1 back and not null.  This commit doesn't get all the way there, but
it does change the convention for EITHER-TEST and gets the system
running and tests passing.

It also eliminates the ENSURE* variant by adding `<opt>` into the
MATCH dialect.  Ultimately, the MATCH dialect and parameter specs in
functions should unify.