Move ACTION_FLAG_XXX out of REBVALs and onto paramlist
Motivated by the desire to free up bits in value cells to implement
the concept of "mirror bytes", this takes the "custom type bits" for
actions and moves them into the paramlist REBSER node, where a large
number of ARRAY_FLAGs were free.

This change has other benefits, such as that a hijacking or other
change to an action can now update these bits...whereas before, they
would be spread out to every value.  A previous workaround was that
the bits were only checked on the canon value in the paramlist--which
was also more inefficient if you had a REBACT* in your hand, to have
to go dereference it to find the first value, then dereference that.