Raise error when pathing bad object field
Rebol2, Red, and R3-Alpha all error when you attempt to use a PATH! to
either pick or poke a field that is not in an object.  Ren-C had been
allowing path picking to return NULL, but as part of fixing a crash
on path poking, this brings it in line with the historical behavior.

Other routines such as SELECT or IN (or FIND?) should be used as
opposed to PATH or PICK when it's not known if the field is in the
object or not.  This should be a bit safer for the average GET-PATH!.