Improve handling of more specialized ordering cases
The code that was introduced to manage walking parameters in their
correct ordering after partial refinement specialization needs to be
used in multiple circumstances.  Notably, when Make_Action() is trying
to cache properties of a function related to its first argument...
and then later if a skippable argument has to be type checked before
its frame is enumerated normally.

This factors out that code to make an enumerator which is passed a hook
that it calls on each unspecialized parameter.  The hook gets two
chances to see the arguments...one in an order that isn't necessarily
the order the parameters will be in, and another in the correct order.
It can use the first pass to make decisions if it can (including how
big to make an array to hold the results), or wait until the second
pass if that is needed.

Using this routine, a "First_Unspecialized_Param()" routine is added,
which will return a quick answer usually...while in pathological cases
it may have to wait for the second pass, but at least the answer will
be correct.  The previous attempt to simulate this in the evaluator
was broken, so this corrects bugs in `<skip>`-able parameters.