Make enfix refinement args honor revocation protocol
There's supposed to be a rule that a refinement argument is null if
(and only if) the refinement says it is not used.  This is enforced in
Finalize_Arg(), which wasn't being called when enfix was handling
missing arguments that turned out to be targeting refinement arg slots.

Also includes a trial of changing the `// null` prompts to `-- null`
in anticipation of a test run of allowing `--` comments.