Use "pseudotypes" to encode parameter classes
In R3-Alpha, a special header flag called "unword" was used to indicate
parameter cells in paramlists.  Their actual type byte was still
reporting an ANY-WORD!.

Ren-C noticed that it was more sensible to think of these as "typesets
with a symbol in them" than as ANY-WORD!s with a bitset in them, as
type checking would presumably get more advanced with time.  So it
twisted this so that they were typesets, with a symbol in the "extra".

However, that led to losing the distinction that the word type gave
for knowing the parameter's "class" (e.g. normal, quoted, refinement)
which had previously come from the word type.  This was addressed by
putting the parameter class in the "type-specific bits" of the header.

The current idea is to phase out type-specific bits, and take ownership
of the header completely for type-independent purposes.  This is
intended to enable "mirror bytes" so that single-cell immutable blocks
Previously the parameter class was put in the CUSTOM byte of a value's
header.