Streamline value flags, removing type-specific bit checks
The concept going forward is that cells are required to get their full
payload done with the three platform-pointers they have, and leave the
header entirely for system purposes (including remembering the type
of the data).  This policy is intended to permit the ambitious idea of
fitting immutable single-element arrays into a cell, thus making a
PATH! that looks like a refinement (e.g. /foo, /1020, or /"foo") take
up no more space than the non-path version.

One benefit of this policy is removing concern about whether it's valid
or not to read a "type specific flag" out of a header.  There was some
complicated debug asserting for this which slowed down the debug build.
It had merit, and was helpful for finding bugs, but the general shift
was to find ways to make the payload express the differences--giving
a performance benefit of not having to keep the payload and the flags
in sync.

There are only a few remaining instances where "type specific bits"
are still used, but they are being phased out.  This goes ahead and
preemptively reduces the complexity of the value flag setting/getting
and clearing routines.