Use 0 size to indicate CFUNC handles
It may be that C-DATA! and C-FUNCTION! replace the idea of a unified
handle, honoring the distinct storage needs.  For the moment there is
one type of HANDLE!, and in order to get it to comply with fitting
its entire information into the payload and extra fields this takes
the HANDLE_FLAG_CFUNC and folds it in to just being 0 length data.