Move missing date timezone and time flags into payload
This continues the new rule that all data types must fit their data in
they payload and extra portions of a cell, and leave the header to
the system to generically manage.  Hence, a special time zone value
and a special time value are used to indicate when there is no time
zone or time in a date.