Change partial specialization to not use header bits
The initial process of partial specialization would leave placeholder
cells in refinement slots while it analyzed what the user had done
with the frame.  These cells made use of some header bits to track
flags on their work in progress.

With the policy of reserving the header bits for "system purposes",
that actually could maybe be accomodated here (it's a system purpose).
But it's possible to do the same thing by using another pseudotype and
a sign bit on an integer in the payload.

This does that, thus eliminating the last usage of the fourth byte from
the header, freeing it up for potential new tricky uses.