Restore "One Expression" Left Normal Enfix Rule
R3-Alpha only evaluated one left-hand side element for infix operators.
Ren-C needed to allow for more evaluation in order to implement various
ideas like ELSE:

     if condition [...] else [...]

The ELSE required some verdict from IF about whether the condition was
true and the branch ran or not.  But if it were limited to only seeing
one block on the left, it couldn't learn anything from the IF.

Argument conventions for the left enfix were decided to be NORMAL,
TIGHT, SOFT QUOTED, and HARD QUOTED.  The only one of these conventions
obeyed by R3-Alpha for infix "OP!s" was the one called TIGHT...which
took one unit of value cell and evaluated it (something like a hard
quoted operator that ran one unit of evaluation).  The question arose
as to how much more evaluation than that a NORMAL argument would do.
Initially it would evaluate as much of the left as it could.

This ran up against a "dark corner", e.g.

    return if condition [...] else [...]

With the interpretation of maximal evaluation, you would get precedence
equivalent to:

    (return if condition [...]) else [...]

The ELSE would never be able to run, because the RETURN was fulfilled,
and would jump out of the routine.  This led to a modificiation of the
rule to be "one expression's worth of evaluation".

This was used for a while, but some concern arose over whether it was
the harder rule to learn in the long run.  Some operators were
introduced to try mitigating concerns about dealing with injecting new
points into the evaluation without using parentheses.  It was theorized
that maybe the operator or mitigating RETURN's problem was an easier
operator to learn.

Switching back to full left completion did not offer the benefits that
were hoped, and raised the need or ugly operators.  This switches
back, with an enhancement to cover one pattern that became common that
previously relied on left completion:

    if condition [10] then x => [print [x]] else [print "boo"]

This was intended to let you use a lambda function as the argument to
the THEN.  However, if the "one expression" rule is in effect, what
happens here is that the ELSE defers the [print [x]] block, but then
acts after the lambda and before the THEN:

    if condition [10] then ((x => [print [x]]) else [print "boo"])

Since lambdas make ACTION!s that are truthy, the ELSE gets thrown
away.  To address this, lambdas were changed to have a TIGHT right-hand
argument, and the rules of deferment do not count tight deferments as
having done the "one deferment" yet.  Hence the first deferment ELSE
counts is the one on lambda itself, allowing it to pin its execution
to after the THEN's completion.