Use DO_FLAG_POST_SWITCH re-entry on hard and soft quotes
Previously, the rules for enfix deferral (one unit of deferral) were
only applied to evaluative (normal) parameters.  It made this work:

    return if condition [...] else [...]

It meant that ELSE would defer acting on the [...] argument until
after the IF processed it, but that it would not defer again.

But quotes did not bother to call into the evaluator for any enfix
processing.  That meant, for instance:

    return lit 1 then (x => [x + 1])

This would return 1, and not 2.  This affects things like the ability
to soft quote branches.  At some point the enfix processing code which
manages deferral has to run.

This attacks the problem somewhat inelegantly, but without adding a lot
of code.  It just factors out the code that is typically used for a
post-switch to do the switch on a regular argument instead of a
deferred one...hence being able to reuse the logic at an arbitrary
point in parameter gathering.