Tweak rule for tightness to only inhibit tight lookahead
In order to get `1 + 2 * 3` to be 9 and not 7, then while the 2 is
being evaluated it has to reject the idea of going forward to the *.
This is accomplished by + marking its right hand side as "tight", so
when the frame recursion happens it is marked with DO_FLAG_NO_LOOKAHEAD

However, inhibiting *all* enfix leads to undesirable properties, like
an error on:

    1 + if false [2] else [3]

If the left enfix normal operator ELSE is treated as if it were another
tight leftward-looking operator, that's read as:

    (1 + if false [2]) else [3]

But this situation would seem to want ELSE to run immediately after
the IF and not after the +.  It's like wanting to run after the IF and
before the RETURN in:

    return if false [2] else [3]

This tweaks the rule so that tight operators only exclude other tight
operators from the lookahead.

Includes bugfix for delivering error messages on undefined path picks.