Update MAP! to correctly handle generalized quoted keys
For maps to use values as keys, they must be ensured as locked.  The
QUOTED! datatype had no handling in the locking mechanism, and it
needed to pass through to its contained data cell.

An oversight in the definition of the KIND_BYTE() primitive led it to
not notice when the kind byte was being extracted from a REBCEL.  That
allowed calling the performance-optimized macros like IS_WORD() on a
cell, which since it was being viewed as a cell would not give the
right answer regarding having a wordlike payload.  Added checking in
the C++ build to only allow KIND_BYTE() on RELVAL* and to force the
use of CELL_KIND() on a REBCEL*.

Cleaned up some code that attempted to be optimized in map insertion,
but was really not optimizing anything much and just making it have its
own errors in determination of strict/lax equality.