Restore comparison operators to enfix tight convention
Rebol2/Red/R3-Alpha had only one type of infix behavior, which was to
evaluate the one item to the left of the "OP!" and use that.

Ren-C added enhancements which found footing in the very useful THEN,
ELSE, and new ALSO operators.  In the interests of overall system
simplicity, it was frequently attempted to make the rules that worked
for these operators be the *only* form of infix that there was.

That didn't ever gel.  But at one point when the new infix operators
were willing to evaluate arbitrarily far on their left, it seemed to
provide more expressivity to leverage that so you could write either
of these expressions without a GROUP!

    if length of block = 1 + 2 [...]
    if 1 + 2 = length of block [...]

The `=` would form a natural split point, completing its left.

As interesting a feature as this is, the idea of "completing the left"
has a troublesome aspect of "how far".

    return a = b

If = doesn't have a notion of where to stop, it would run RETURN A.

The unpredictable nature of how much material will be collapsed also
complicates the evaluator.  In doing so, it complicates the interface
for variadic functions to where they cannot describe their behavior
simply in terms of taking parameters one at a time...but to match the
sophistication of the evaluator they must sneakily fill slots they
may decide to revisit later.

In order to lay down some final rules on the design, the properties
that power THEN, ELSE, and ALSO are being tuned to that specific need.
And the default sense of infix operators is returning to being that
of the historical Rebol evaluator.  This switches comparison operators
back to their previous nature.