Make AND/OR/XOR tight enfix, accept QUOTED! right
In the interests of pinning down a complete and understandable story
of enfix for Beta/One, this makes the conditional boolean logic
operators tight enfix.  This allows the focus of the new feature to
support THEN/ELSE/ALSO to be tailored specifically to those cases, and
to hopefully simplify the evaluator in a way that will streamline
performance and parity with a simple model for variadic access.

Left enfix completion for AND and OR had originally been motivated by
there being LOGIC!-bearing routines like PARSE's return result which
could not be accomodated by ELSE.  That means a big part of this commit
is switching ORs to ELSEs in cases that had used that pattern.

There were relatively few left enfix completing AND and OR that were
designed to be used like THEN and ELSE.  These could be handled on a
case-by-case basis...either by scrapping the idea or by using a GROUP!
around the expression it was trying to tie to.

The vast majority of other cases were just taking things that had more
than a single argument to the left of an AND or OR, and either putting
that into a GROUP! or deciding to rewrite the expression.  There are
a lot of possibilities, such as transforming to ALL and ANY, which can
cut down on the number of blocks and symbols used in total while
trading off for vertical space:

    if something something or [something something] [
         fail [...]
    ]
    =>
    any [
        something something
        something something
        fail [...]
    ]

Also adds the ability for AND and OR to take a QUOTED! word or path as
the right hand side, as a synonym for BLOCK!.