Add validation and comments to frame fill of RETURN + LOCAL
There were some ?? comments in the evaluator seeming to ask about the
invariants of the state of the f->special field when filling return
and local frame variables.  There's actually not a lot you can assert
about them; this adds what little there is and explains why it's so
little you can take for granted.

Also takes care of a stray IF in a parse rule, now replaced by general
handling of LOGIC! in rules.