Avoid overwrite when RETURN is used in invisibles
The way that invisibles work is to leave the contents in a frame's
output cell, instead running the evaluation of the user's code result
into a dummy cell.  However, if there is a throw generated that is
going to bubble up, it must be moved into the output cell to do so...
overwriting any result.

That's is usually the right thing to do, *unless* the throw corresponds
to something that specific invisible's frame intends to catch.  In
which case, it just threw away the information it was trying to
preserve.

This is a stopgap measure which just catches RETURNs in the invisible's
dispatcher.  A more elegant solution should be able to avoid repeating
the code--since there is already another more complex case (support
for REDO in tail call recursions).