Simplify deferred enfix to just a frame flag
A complex method of re-entering the evaluator was devised in order
to allow deferred enfix to appear to act as either greedy or not,
without a-priori knowledge of whether it was on the last parameter:

    if 1 and 2 [...]
    if negate 1 and 2 [...]

It managed to do this by keeping track of where in the parameter list
it was when it did the last deferral, and if it ever found another
unspecialized parameter in the frame it would belatedly go back and
run as if it were tight...otherwise it would belatedly type check what
it had originally put in the slot.

This involved extra bookkeeping.  But more importantly, it made the
evaluator complicated...complicated enough that things that were
more simple (such as a sequence of variadic argument takes) could not
do the same magic without an a-priori announcement of when their last
argument take was.

Given that decisions need to be made for Beta/One, the evidence was in
favor of eliminating this magic.  It offered too little benefit, and
caused too much "nuance" in the evaluator...leading to breakage of
other cool features.  As an advantage of eliminating it, what had
previously required multiple frame fields and a delicate system of
management could not be collapsed into a single frame flag: are we in
a deferred enfix mode or not.

This gave a welcome simplification to the evaluator.