Export MAKE FRAME! from a VARARGS!, restore MATCH experiment
There was a MATCH experiment that had broken the ability of uses of
the non-experimental MATCH to be combined with THEN or ELSE.  This
was addressed by the uniform variadic policy to deferred enfix, by
simplifying the evaluator so it couldn't use its "inside knowledge"
to accomplish results more complex than the variadic, instead leading
the variadic to see a THEN or ELSE as being an end of input looking
toward the right.

Hence the experiment is restored here.  This also exports the mechanic
by which the effect is achieved, so that a userspace function could
achieve the same thing.