Access bits through uint32_fast_t union field
Although Reb_Headers are the size of a pointer (e.g. 64 bits on 64-bit
machines), it might be that 32-bit integers are processed faster,
and only 32 bits are used.  We do want to reserve the size of a full
platform pointer (uintptr_t), and alignment would give it to us even
if we didn't ask for it.  But it's possible to make the .bits union
field that accesses are currently done through a uint_fast32_t.