Move enfix flags, hold flag, barrier hit to shared "feed"
While processing a stream of source at one level of depth (either an
array or a varargs for now), many frames can be constructed.  e.g.

    multiply 1 + 2 3 + 4

There is a frame for the multiply, which then spawns a subframe that
advances the same array to run add.  These advances must be seen by
the multiply in order to know where to start the next argument
gathering.

This is managed by what was previously called the frame's "source", but
is now the frame's "feed".  It is a shared structure that has its
index advanced in a way that all frames will see the advancement.

In order to implement the new scheme for managing enfix lookahead and
deferment, it's much easier to put flags for it in the feed instead
of trying to keep them in sync as flags on the frame.  This is also
a good fit for the flag for whether a feed took a hold or not, or
whether a barrier was hit.

Besides providing the key to implementing the finalized version of
enfix logic, this also frees up some more frame flags.