Allow ACTION OF on heap-based FRAME!s
When you do MAKE FRAME! explicitly, it does not have a runtime REBFRM*
associated with it unless it is executing.  The reflectors like
PARAMETERS OF were all assuming you could get a REBFRM* for it, but
you don't need that to get the parameter list...as it comes from the
keys of the frame.  This breaks out the PARAMETERS OF reflector as
a separate accessor.