RESKINNED technique for rewriting parameter conventions
This is a proof of concept for changing the parameter conventions of
functions--either the types they accept, or the parameter class and
its quotedness.  The dialect is very preliminary and definitely needs
more design work, but it covers the basic operations.

You can #change a parameter's category, for instance to make a variant
of APPEND which appended its argument literally:

    append-q: reskinned [#change :value] :append
    [a b c d] = append-q mutable [a b c] d

You can expand types that a function accepts or returns with #add.

    foo: func [x [integer!]] [x]
    skin: reskinned [x #add [text!]] (adapt :foo [x: to integer! x])

    >> skin "10"
    == 10

    >> skin 10
    == 10

Similarly you can remove them with #remove.  If you don't include #add
or #remove it is assumed you want to overwrite with a type block
completely.

The type block may be omitted if you are using #change to change the
parameter convention.

Any reskinning that expands argument types (or restricts return types)
requires the injection of a new dispatcher.  Because this dispatcher
must have a way to guarantee it can't leak unexpected type bits into
natives (which would cause crashes), the only functions that may be
reskinned in an expansive way are those that do a parameter check
after usermode code runs...so EXPAND and ADAPT.

Narrowing parameter cases, or broadening return cases, do not require
new checks on top of what the target action already did.