Give FAIL a skippable lit-word! argument to "blame"
Now that `<skip>` parameters are a confirmed Beta/One feature, one of
the desired usages was to allow FAIL to supply a place to blame.  It
was desired to do this with @words, e.g. `fail @location [...]`, but
for now a quoted WORD! will do, as `fail 'location [...]`

It works for parameter values resident in frames, but also just for
plain paths and values.  If the word is a parameter, then it will
also inform the WHERE of the error to implicate the callsite (instead
of just showing the point that called FAIL)