Remove IF-NOT, WHILE-NOT, UNTIL-NOT
IF-NOT was created as an "optimized" replacement for UNLESS, but which
also tried to pick up the slack for the apparent "precedence change"
created by making conditional operators complete their left hand side.

The decision to revert those operators to their conventional behavior
and reserve the new deferred precedence for THEN, ELSE, ALSO/etc.
means that IF NOT is now a synonym for what was UNLESS.

WHILE-NOT was created for a similar dualism of thinking it would be
faster to fold the NOT into the native, plus to try and workaround the
new "precedence" rule.  UNTIL-NOT was created just because it seemed
like a thing to do since WHILE had an inverted version.

None got used, and now with NOT returning to its old rules there's
*really* very little reason to use them.  The focus can be on just
making userspace calls to NOT faster.